requirements.txt
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
python-chess>=1.999
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
rich>=13.0.0
pyyaml>=6.0
